convert canvas prepare to typescript
